authorization

+++ FastApi/Project-2 (Todo App)/routers/todos.py
@@ -5,6 +5,7 @@ from fastapi import APIRouter, Depends, HTTPException, Path
 from models import Todos
 from database import session_local
 from starlette import status
+from .auth import get_current_user
 
 router = APIRouter()
 
@@ -16,6 +17,7 @@ def get_db():
         db.close()
 
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     title: str = Field(min_length=3)
@@ -24,26 +26,32 @@ class TodoRequest(BaseModel):
     complete: bool
 
 @router.get("/")
-async def read_all(db : db_dependency):
-    return db.query(Todos).all()
+async def read_all(user: user_dependency, db : db_dependency):
+    return db.query(Todos).filter(Todos.owner_id == user.get("id")).all()
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo(db : db_dependency,todo_id: int = Path(gt=0)):
-    todo = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user: user_dependency, db : db_dependency,todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed!")
+    todo = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).first()
     if todo is not None:
         return todo
     raise HTTPException(status_code=404, detail="Todo not found")
 
 @router.post("/create_todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(db : db_dependency, todo_request: TodoRequest):
-    todo = Todos(**todo_request.model_dump())
+async def create_todo(user: user_dependency, db : db_dependency, todo_request: TodoRequest):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed!")
+    todo = Todos(**todo_request.model_dump(), owner_id=user.get("id"))
 
     db.add(todo)
     db.commit()
 
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db : db_dependency, todo_request: TodoRequest, todo_id: int = Path(gt=0)):
-    todo = db.query(Todos).filter(Todos.id == todo_id).first()
+async def update_todo(user: user_dependency, db : db_dependency, todo_request: TodoRequest, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed!")
+    todo = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).first()
     if todo is None:
         raise HTTPException(status_code=404, detail="Todo not found")
     todo.title = todo_request.title
@@ -54,9 +62,11 @@ async def update_todo(db : db_dependency, todo_request: TodoRequest, todo_id: in
     db.commit()
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db : db_dependency, todo_id: int = Path(gt=0)):
-    todo = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user: user_dependency, db : db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail="Authentication failed!")
+    todo = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).first()
     if todo is None:
         raise HTTPException(status_code=404, detail="Todo not found")
-    db.query(Todos).filter(Todos.id == todo_id).delete()
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).delete()
     db.commit()
